Create sys-essentials-aggregate-nics.py
Adding script for configuring network aggregation in the Steam Deck. The idea is that this should allow for uninterrupted access to the game library via NFS when transitioning seamlessly between docked and undocked modes. Assumes the ethernet interface is properly restarted by the kernel when redocking. However, it should default to the wireless interface if we lose the network interface when docked.